Updates to Game Schema to better match API

+++ PlayTogether/Server/Models/GameGenre.cs
@@ -12,7 +12,7 @@ namespace PlayTogether.Server.Models
         public int Id { get; set; }
 
         [Required]
-        public int ApiId { get; set; }
+        public long ApiId { get; set; }
 
         [Required]
         public string Name { get; set; }
