+++ PlayTogether/Server/Models/GamingPlatform.cs
@@ -11,7 +11,7 @@ namespace PlayTogether.Server.Models
         public int Id { get; set; }
 
         [Required]
-        public int ApiId { get; set; }
+        public long ApiId { get; set; }
 
         public string Abbreviation { get; set; }
